refactor: rename `finder`
