+++ src/awkward/_nplikes/jax.py
@@ -5,7 +5,7 @@ import numpy
 
 import awkward as ak
 from awkward._nplikes.array_module import ArrayModuleNumpyLike
-from awkward._nplikes.finder import register_nplike
+from awkward._nplikes.dispatch import register_nplike
 from awkward._nplikes.numpylike import ArrayLike
 from awkward._typing import Final
