+++ src/awkward/_nplikes/numpy.py
@@ -5,14 +5,12 @@ import numpy
 
 import awkward as ak
 from awkward._nplikes.array_module import ArrayModuleNumpyLike
-from awkward._nplikes.finder import register_nplike
 from awkward._nplikes.numpylike import ArrayLike, NumpyMetadata
 from awkward._typing import Final, Literal
 
 np = NumpyMetadata.instance()
 
 
-@register_nplike
 class Numpy(ArrayModuleNumpyLike):
     is_eager: Final = True
 
@@ -32,7 +30,7 @@ class Numpy(ArrayModuleNumpyLike):
         return self._module.ndarray
 
     @classmethod
-    def is_own_array_type(cls, obj: type) -> bool:
+    def is_own_array(cls, obj) -> bool:
         """
         Args:
             obj: object to test
@@ -40,7 +38,7 @@ class Numpy(ArrayModuleNumpyLike):
         Return `True` if the given object is a numpy buffer, otherwise `False`.
 
         """
-        return issubclass(obj, numpy.ndarray)
+        return isinstance(obj, numpy.ndarray)
 
     def is_c_contiguous(self, x: ArrayLike) -> bool:
         return x.flags["C_CONTIGUOUS"]
