+++ src/awkward/_nplikes/typetracer.py
@@ -7,7 +7,7 @@ import numpy
 from numpy.lib.mixins import NDArrayOperatorsMixin
 
 import awkward as ak
-from awkward._nplikes.finder import register_nplike
+from awkward._nplikes.dispatch import register_nplike
 from awkward._nplikes.numpylike import ArrayLike, IndexType, NumpyLike, NumpyMetadata
 from awkward._nplikes.shape import ShapeItem, unknown_length
 from awkward._regularize import is_integer, is_non_string_like_sequence
